requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow-cpu==2.13.0
